requirements.txt
jieba
sklearn
line-bot-sdk
Flask
scipy
gensim
gunicorn
numpy
torch